feat: 型定義ファイルの作成
